feat!: split heavy-dep wrapper transports into sub-modules
Five vendor wrappers (zerolog, zap, logrus, phuslu, charmlog) are now
their own Go modules so consumers only pay for the SDKs they actually
import. The main go.loglayer.dev module's direct deps drop from 8
(fatih/color + 5 vendor SDKs + goccy + goleak) to 3 (fatih/color, goccy,
goleak); a structured-only consumer goes from 50 lines of go.sum to 12.

Each wrapper has its own go.mod with a `replace go.loglayer.dev => ../..`
directive for development. The cross-comparison Direct_X vs Wrapped_X
benchmarks moved from the root bench_test.go into each wrapper's own
bench_test.go so the main module no longer needs to import any vendor
SDK to build its tests. Render_* and Loglayer_* benches stay in the
root since they only use first-party renderers.

Updated:
- scripts/foreach-module.sh: ALL_MODULES, SHIPPED_MODULES, and the
  hardcoded test op now include the new modules.
- .release-please-config.json / .release-please-manifest.json: each new
  sub-module gets its own release pipeline with prefixed tags
  (transports/zerolog/v1.0.0, etc.).
- AGENTS.md "Mostly single Go module" → "Multi-module layout".
- CHANGELOG.md: notes the per-wrapper module split.
- Each wrapper's transport doc page gets a `:::info Separate module`
  block explaining the dep-graph isolation.
- transports/otellog test that previously asserted JSON-roundtrip's
  int->float64 widening now asserts AsInt64 (the reflect walker
  preserves Go int).

Stays in main: framework core, slog wrapper (stdlib), renderers
(blank/console/pretty/structured/testing), network transports
(http/datadog), integrations/loghttp, and the utility/plugin packages
that only depend on stdlib + goccy/go-json.

+++ transports/otellog/otellog_test.go
@@ -131,8 +131,7 @@ func TestStructMetadataNested(t *testing.T) {
 		t.Fatalf("metadata should be Map, got %v", meta.Kind())
 	}
 	got := kvMap(meta.AsMap())
-	// JSON roundtrip turns Go ints into float64.
-	if got["id"].AsFloat64() != 7 || got["name"].AsString() != "Alice" {
+	if got["id"].AsInt64() != 7 || got["name"].AsString() != "Alice" {
 		t.Errorf("nested fields: got %v", got)
 	}
 }
